@@ -26,8 +26,8 @@ import org.springframework.web.servlet.HandlerMapping
 import spock.lang.Specification
 import spock.lang.Unroll
 
-import javax.servlet.http.HttpServletRequest
-import javax.servlet.http.HttpServletResponse
+import jakarta.servlet.http.HttpServletRequest
+import jakarta.servlet.http.HttpServletResponse
 import java.rmi.AccessException
 import java.rmi.AlreadyBoundException
 import java.rmi.NotBoundException
